+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 <h2>v1.0.0-alpha32</h2>
 
+<h3>Pipelines</h3>
+* Rewrote to include a fluent syntax, eg `(new Pipeline($container))->send("foo")->through($stages)->then($callback)->execute()`
+
 <h3>Validation</h3>
 * Added `Opulence\Validation` library
 * Added `Opulence\Framework\Bootstrappers\Validation\ValidatorBootstrapper`